+++ src/main/java/salute/oneshot/domain/payment/controller/PaymentController.java
@@ -2,14 +2,15 @@ package salute.oneshot.domain.payment.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import salute.oneshot.domain.common.dto.success.ApiResponse;
 import salute.oneshot.domain.common.dto.success.ApiResponseConst;
 import salute.oneshot.domain.payment.dto.request.ConfirmPaymentRequestDto;
 import salute.oneshot.domain.payment.dto.response.ConfirmPaymentResponseDto;
-import salute.oneshot.domain.payment.dto.response.PaymentResponseDto;
 import salute.oneshot.domain.payment.dto.service.ConfirmPaymentSDto;
-import salute.oneshot.domain.payment.entity.PaymentStatus;
 import salute.oneshot.domain.payment.service.PaymentService;
 
 @RestController
