+++ src/main/java/salute/oneshot/domain/payment/dto/feign/TossCancelPaymentRequestDto.java
@@ -1,0 +1,18 @@
+package salute.oneshot.domain.payment.dto.feign;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class TossCancelPaymentRequestDto {
+
+    private final String cancelReason;
+
+    public static TossCancelPaymentRequestDto of(String cancelReason) {
+        return new TossCancelPaymentRequestDto(
+                cancelReason
+        );
+    }
+}
