+++ src/main/java/salute/oneshot/domain/payment/dto/feign/TossConfirmPaymentRequestDto.java
@@ -1,0 +1,22 @@
+package salute.oneshot.domain.payment.dto.feign;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class TossConfirmPaymentRequestDto {
+
+    private final String orderId;
+    private final String amount;
+    private final String paymentKey;
+
+    public static TossConfirmPaymentRequestDto of(String orderId, String amount, String paymentKey) {
+        return new TossConfirmPaymentRequestDto(
+                orderId,
+                amount,
+                paymentKey
+        );
+    }
+}
