+++ src/main/java/salute/oneshot/domain/payment/util/PaymentClient.java
@@ -2,17 +2,25 @@ package salute.oneshot.domain.payment.util;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import salute.oneshot.domain.payment.dto.request.ConfirmPaymentRequestDto;
-import salute.oneshot.domain.payment.dto.response.ConfirmPaymentResponseDto;
-import salute.oneshot.domain.payment.dto.service.ConfirmPaymentSDto;
+import salute.oneshot.domain.payment.dto.feign.TossCancelPaymentRequestDto;
+import salute.oneshot.domain.payment.dto.feign.TossConfirmPaymentRequestDto;
+import salute.oneshot.domain.payment.dto.feign.TossPaymentResponseDto;
 import salute.oneshot.global.config.FeignConfig;
 
 //@FeignClient(name = "paymentClient", url = "${payment.base-url}", configuration = PaymentFeignConfig.class)
 @FeignClient(name = "paymentClient", url = "https://api.tosspayments.com/v1/payments", configuration = FeignConfig.class)
 public interface PaymentClient {
+
     @PostMapping(value = "/confirm", consumes = MediaType.APPLICATION_JSON_VALUE)
-    ConfirmPaymentResponseDto confirmPayment(@RequestBody ConfirmPaymentSDto sdto);
+    TossPaymentResponseDto confirmPayment(@RequestBody TossConfirmPaymentRequestDto requestDto);
+
+    @PostMapping(value = "/order/{orderId}", consumes = MediaType.APPLICATION_JSON_VALUE)
+    TossPaymentResponseDto getPaymentByOrderId(@PathVariable Long orderId);
+
+    @PostMapping(value = "/cancel", consumes = MediaType.APPLICATION_JSON_VALUE)
+    TossPaymentResponseDto cancelPayment(@RequestBody TossCancelPaymentRequestDto requestDto);
 
 }
