+++ src/main/java/salute/oneshot/global/config/FeignConfig.java
@@ -6,6 +6,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import salute.oneshot.domain.payment.util.PaymentAuthInterceptor;
 import salute.oneshot.domain.payment.util.PaymentClient;
+import salute.oneshot.domain.payment.util.PaymentErrorDecoder;
 
 import java.util.concurrent.TimeUnit;
 
@@ -16,6 +17,11 @@ public class FeignConfig {
     @Bean
     public Request.Options requestOptions() {
         return new Request.Options(2, TimeUnit.SECONDS, 60, TimeUnit.SECONDS, true);
+    }
+
+    @Bean
+    public PaymentErrorDecoder paymentErrorDecoder() {
+        return new PaymentErrorDecoder();
     }
 
     @Bean
