+++ src/main/resources/templates/payment/fail.html
@@ -17,6 +17,8 @@
     const codeElement = document.getElementById("code");
     const messageElement = document.getElementById("message");
 
-    codeElement.textContent = "에러코드: " + urlParams.get("httpStatus");
-    messageElement.textContent = "실패 사유: " + urlParams.get("errorMessage");
+    // codeElement.textContent = "에러코드: " + urlParams.get("httpStatus");
+    // messageElement.textContent = "실패 사유: " + urlParams.get("errorMessage");
+    codeElement.textContent = "에러코드: " + urlParams.get("code");
+    messageElement.textContent = "실패 사유: " + urlParams.get("message");
 </script>
